fix: Add version constraints to the providers (#74)
By not having version constraints (as opposed to pinning provider
versions) in a module, it can force a provider upgrade (and lockfile
change) or cause an `init` failure to downstream consumers.

I have looked through previous versions of each provider and determined
which version is the earliest possible version that could be used with
this module.